Working on index, footer, and harmonizing pages
The index page is a terrible hack, and hopefully I can figure out the templating system so that I don't have to use JS just to get the stylesheet working. But it works for now, and I can resolve the other issue later. I also fixed up the privacy page, but will need to talk to MJ about that.